Updated space_map

+++ kernel/x86_64/memory/space.c
@@ -41,7 +41,7 @@ uintptr_t space_map(Space* space, uintptr_t phys_addr, size_t size, uintptr_t vi
     lock(&space->lock);
 
     // map all the required pages
-    for (uint32_t i = 0; i < size; i++) {
+    for (size_t i = 0; i < size; i++) {
         vmm_map_page(space->page_table, phys_addr + (i*PAGE_SIZE), virt_addr + (i*PAGE_SIZE), true, true);
     }
